Have consistent code generation task names across methods and iface generation
